some progress on the live mapper

+++ src/index.js
@@ -5,7 +5,8 @@ import YjsSync from './js/misc/YjsSync.js';
 import SyncMeta from 'syncmeta-plugin';
 import Main from './js/Main.js';
 import TagRegistry from './js/tags/TagRegistry.js';
-import { mxEvent } from './js/misc/mxExport.js';
+import { mxEvent, mxCodec, mxUtils } from './js/misc/mxExport.js';
+import CONST from './js/misc/Constants.js';
 $(function () {
     YjsSync('yireframetesting').done(function (y) {
         var vls = y.share.data.get('metamodel');
@@ -19,10 +20,14 @@ $(function () {
         //Important load a vls before calling Main
         var editor = Main();
         GoogleLogin();
+
+        //The live mapper starts here
         SyncMeta.init(y);
         var model = y.share.data.get('model');
         var widgetNodeId = null;
         var hasChildMap = {};
+        var nodeCreateMap = {};
+        var mxCellsAddedFlag = true, mxCellsRemoveFlag = true;
         if (model) {
             for (var key in model.nodes) {
                 if (model.nodes.hasOwnProperty(key)) {
@@ -51,12 +56,44 @@ $(function () {
 
         SyncMeta.onNodeAdd(function (event) {
             mxLog.writeln('Node was created in SyncMeta: ' + JSON.stringify(event));
+            nodeCreateMap[event.id] = event;
         });
-
         SyncMeta.onEdgeAdd(function (event) {
             mxLog.writeln('Edge was created in SyncMeta: ' + JSON.stringify(event));
-            if (event.source === widgetNodeId)
-                mxLog.writeln('Node connected to the Widget');
+            var cell = editor.graph.model.getCell(event.target);
+            if (cell) return;
+            if (event.source === widgetNodeId && event.type === 'Widget to HTML Element') {
+                var ymap = y.share.nodes.get(event.target);
+                if (ymap) {
+                    var attr = ymap.get(event.target + '[type]');
+                    var type;
+                    if (attr)
+                        type = attr.value;
+                    else
+                        type = 'input';
+
+                    var UIControl = editor.getUIComponentFromHTMLName(type);
+                    if (UIControl) {
+                        var encoder = new mxCodec();
+                        var uiControl = new UIControl();
+                        var result = encoder.encode(uiControl);
+
+                        var xml = mxUtils.getXml(result);
+                        mxCellsAddedFlag = false;
+                        //Create Node
+                        y.share.action.set(mxEvent.ADD_VERTEX, {
+                            userId: y.db.userId,
+                            id: event.target,
+                            data: xml,
+                            parent: null
+                        });
+                        if (editor.graph.getDefaultParent().children.length > 1) {
+                            y.share.action.set(CONST.ACTIONS.SHARED.APPLY_LAYOUT, { userId: y.db.userId, cellId: null });
+                        }
+                    }
+
+                }
+            }
         });
         SyncMeta.onNodeDelete(function (event) {
             mxLog.writeln('Node deleted from Widget: ' + JSON.stringify(event));
@@ -69,13 +106,51 @@ $(function () {
         SyncMeta.onEdgeDelete(function (event) {
             mxLog.writeln('Edge deleted from the Widget: ' + JSON.stringify(event));
             var cell = editor.graph.model.getCell(event.target);
-            y.share.action.set(mxEvent.REMOVE, {
-                userId: y.db.userId,
-                cells: [cell.id]
-            });
+            if (event.type === 'Widget to HTML Element') {
+                y.share.action.set(mxEvent.REMOVE, {
+                    userId: y.db.userId,
+                    cells: [cell.id]
+                });
+            }
+        });
+        SyncMeta.onNodeAttributeChange(function (value, entity, entityValueId) {
+            var cell = editor.graph.model.getCell(entity);
+            var UIControl = editor.getUIComponentFromHTMLName(value);
+            if (!value || !cell || cell.constructor.HTML_NODE_NAME === value || entityValueId.indexOf('[type]') == -1) return;
+            if (UIControl) {
+                mxCellsRemoveFlag = false;
+                y.share.action.set(mxEvent.REMOVE, {
+                    userId: y.db.userId,
+                    cells: [entity]
+                });
+
+                var uiControl = new UIControl();
+                uiControl.geometry.x = cell.geometry.x;
+                uiControl.geometry.y = cell.geometry.y;
+                var encoder = new mxCodec();
+                var result = encoder.encode(uiControl);
+
+                var xml = mxUtils.getXml(result);
+                mxCellsAddedFlag = false;
+                y.share.action.set(mxEvent.ADD_VERTEX, {
+                    userId: y.db.userId,
+                    id: entity,
+                    data: xml,
+                    parent: cell.parent
+                });
+            } else {
+                y.share.action.set(mxEvent.REMOVE, {
+                    userId: y.db.userId,
+                    cells: [entity]
+                });
+            }
         });
 
         editor.graph.addListener(mxEvent.CELLS_ADDED, function (graph, event) {
+            if (!mxCellsAddedFlag) {
+                mxCellsAddedFlag = true;
+                return;
+            }
             var cells = event.getProperty('cells');
             var absolute = event.getProperty('absolute');
             var parent = event.getProperty('parent');
@@ -87,8 +162,10 @@ $(function () {
                     setTimeout(function () {
                         SyncMeta.setAttributeValue(cell.id, 'type', cell.constructor.HTML_NODE_NAME);
                         SyncMeta.createEdge('Widget to HTML Element', widgetNodeId, cell.id);
-                        if (parent.id != '1')
-                            SyncMeta.createEdge('hasChild', parent.id, cell.id);
+                        if (parent.id != '1') {
+                            var edgeId = SyncMeta.createEdge('hasChild', parent.id, cell.id);
+                            hasChildMap[cell.id] = edgeId;
+                        }
 
                     }, 500);
                 }
@@ -111,7 +188,11 @@ $(function () {
                 SyncMeta.applyLayout();
             }, 1000);
         });
-        editor.graph.addListener(mxEvent.REMOVE_CELLS, function (graph, event) {
+        editor.graph.addListener(mxEvent.REMOVE_CELLS, function (graph,  event) {
+            if (!mxCellsRemoveFlag) {
+                mxCellsRemoveFlag = true;
+                return;
+            }
             var cells = event.getProperty('cells');
             for (var i = 0; i < cells.length; i++) {
                 SyncMeta.deleteNode(cells[i].id);
@@ -120,7 +201,6 @@ $(function () {
                 }, 500);
             }
         });
-
         editor.graph.addListener(mxEvent.GROUP_CELLS, function (graph, event) {
             var test = true;
         });
